% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q2c.R
\name{q2c}
\alias{q2c}
\title{Qualtrics (wide) to Conjoint (long)}
\usage{
q2c(data, prefix = "F", id, outcome, covariates, type = "choice")
}
\arguments{
\item{data}{a \code{data.frame} object, \code{tibble} object, file path (\code{.csv}) or url (\code{.csv})}

\item{prefix}{a prefix of attributes and levels (default is \code{"F"})}

\item{id}{Respondents' ID}

\item{outcome}{a vector of outcome variables.}

\item{covariates}{a vector of covariates.}

\item{type}{type of outcome variable (\code{"choice"} (default), \code{"rating"}, or \code{"rank"})}
}
\value{
A \code{tibble} object.
}
\description{
Qualtrics (wide) to Conjoint (long)
}
\details{
\describe{
\item{\code{data}}{データをパスで指定する場合、データはQualtrics形式である必要があります。}
}
}
