% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q2c.R
\name{q2c}
\alias{q2c}
\title{Qualtrics to Wide Data Conjoint}
\usage{
q2c(data, prefix = "F", id, outcome, covariates, type = "choice")
}
\arguments{
\item{data}{読み込み済みのデータフレーム、またはデータのパス}

\item{prefix}{属性と水準列名の接頭詞（既定値は\code{"F"}）}

\item{id}{回答者のID}

\item{outcome}{応答変数}

\item{covariates}{共変量}

\item{type}{応答変数のタイプ (\code{"choice"} (default), \code{"rating"}, or \code{"rank"})}
}
\value{
A \code{tibble} object.
}
\description{
Qualtrics to Wide Data Conjoint
}
\details{
\describe{
\item{\code{data}}{データをパスで指定する場合、データはQualtrics形式である必要があります。}
}
}
